+++ TriSprint/Utils/FormatDisplay.swift
@@ -48,7 +48,7 @@ struct FormatDisplay {
     }
     
     
-    static func time(_ seconds: Int) -> String {
+    static func time(_ seconds: Double) -> String {
         let formatter = DateComponentsFormatter()
         formatter.allowedUnits = [.hour, .minute, .second]
         formatter.unitsStyle = .positional
@@ -56,49 +56,49 @@ struct FormatDisplay {
         return formatter.string(from: TimeInterval(seconds))!
     }
     
-    static func pacePerMile(distance: Measurement<UnitLength>, seconds: Int, outputUnit: UnitSpeed) -> String {
+    static func pacePerMile(distance: Measurement<UnitLength>, seconds: Double, outputUnit: UnitSpeed) -> String {
         let formatter = MeasurementFormatter()
         formatter.unitOptions = [.providedUnit]
         let distanceMiles = distance.value/1609
-        let speedMagnitude = seconds != 0 ? Double(seconds / 60) / distanceMiles : 0
+        let speedMagnitude = seconds != 0 ? (seconds / 60) / distanceMiles : 0
         let speed = Measurement(value: speedMagnitude, unit: UnitSpeed.minutesPerMile)
         return formatter.string(from: speed.converted(to: outputUnit))
     }
     
-    static func pacePerKm(distance: Measurement<UnitLength>, seconds: Int, outputUnit: UnitSpeed) -> String {
+    static func pacePerKm(distance: Measurement<UnitLength>, seconds: Double, outputUnit: UnitSpeed) -> String {
         let formatter = MeasurementFormatter()
         formatter.unitOptions = [.providedUnit]
         let distanceKm = distance.value / 1000
-        let speedMagnitude = seconds != 0 ? Double(seconds / 60) / distanceKm : 0
+        let speedMagnitude = seconds != 0 ? (seconds / 60) / distanceKm : 0
         let speed = Measurement(value: speedMagnitude, unit: UnitSpeed.minutesPerKilometer)
         return formatter.string(from: speed.converted(to: outputUnit))
     }
     
-    static func speedMph(distance: Double, seconds: Int16, outputUnit: UnitSpeed) -> String {
+    static func speedMph(distance: Double, seconds: Double, outputUnit: UnitSpeed) -> String {
         let formatter = MeasurementFormatter()
         formatter.unitOptions = [.providedUnit]
-        let hourTime = Double(seconds) / 3600
+        let hourTime = seconds / 3600
         let kmDistance = distance / 1000
         let mileDistance = kmDistance / 1.609
-        let speedMagnitude = seconds != 0 ? mileDistance / Double(hourTime) : 0
+        let speedMagnitude = seconds != 0 ? mileDistance / hourTime : 0
         let speed = Measurement(value: speedMagnitude, unit: UnitSpeed.milesPerHour)
         return formatter.string(from: speed)
     }
     
-    static func speedKmph(distance: Double, seconds: Int16, outputUnit: UnitSpeed) -> String {
+    static func speedKmph(distance: Double, seconds: Double, outputUnit: UnitSpeed) -> String {
         let formatter = MeasurementFormatter()
         formatter.unitOptions = [.providedUnit]
-        let hourTime = Double(seconds) / 3600
+        let hourTime = (seconds / 3600)
         let kmDistance = distance / 1000
-        let speedMagnitude = seconds != 0 ? kmDistance / Double(hourTime) : 0
+        let speedMagnitude = seconds != 0 ? kmDistance / hourTime : 0
         let speed = Measurement(value: speedMagnitude, unit: UnitSpeed.kilometersPerHour)
         return formatter.string(from: speed)
     }
     
-    static func pacePerMileDble(distance: Double, seconds: Int16, outputUnit: UnitSpeed) -> String {
+    static func pacePerMileDble(distance: Double, seconds: Double, outputUnit: UnitSpeed) -> String {
         let formatter = MeasurementFormatter()
         formatter.unitOptions = [.providedUnit]
-        let minutes = Double(seconds / 60)
+        let minutes = (seconds / 60)
         
         let kmDistance = distance / 1000
         let mileDistance = kmDistance / 1.609
@@ -106,10 +106,10 @@ struct FormatDisplay {
         let speed = Measurement(value: speedMagnitude, unit: UnitSpeed.minutesPerMile)
         return formatter.string(from: speed)
     }
-    static func pacePerKmDble(distance: Double, seconds: Int16, outputUnit: UnitSpeed) -> String {
+    static func pacePerKmDble(distance: Double, seconds: Double, outputUnit: UnitSpeed) -> String {
         let formatter = MeasurementFormatter()
         formatter.unitOptions = [.providedUnit]
-        let minutes = Double(seconds / 60)
+        let minutes = (seconds / 60)
         print("Nige: km secs = \(seconds),  minutes = \(minutes)")
         let kmDistance = distance / 1000
         let speedMagnitude = seconds != 0 ? minutes / kmDistance : 0
